Move traceStatusTranslation to utils file
